+++ golang/kurtosis_lambda_docker_api/kurtosis_lambda_docker_api.go
@@ -6,10 +6,6 @@
 package kurtosis_lambda_docker_api
 
 const (
-	// A string defining the loglevel that the Lambda should output
-	// This string should be meaningful to the user, since Kurtosis won't know what logging framework they use
-	LogLevelEnvVar = "LOG_LEVEL"
-
 	// IP:port of the Kurtosis API container
 	ApiContainerSocketEnvVar = "API_CONTAINER_SOCKET"