+++ golang/lib/execution/lambda_executor.go
@@ -6,6 +6,10 @@
 package execution
 
 import (
+	"github.com/kurtosis-tech/kurtosis-client/golang/kurtosis_core_rpc_api_bindings"
+	"github.com/kurtosis-tech/kurtosis-client/golang/lib/networks"
+	"github.com/kurtosis-tech/kurtosis-client/golang/lib/services"
+	"github.com/kurtosis-tech/kurtosis-lambda-api-lib/golang/kurtosis_lambda_docker_api"
 	"github.com/kurtosis-tech/kurtosis-lambda-api-lib/golang/kurtosis_lambda_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-lambda-api-lib/golang/kurtosis_lambda_rpc_api_consts"
 	"github.com/kurtosis-tech/minimal-grpc-server/server"
@@ -20,42 +24,38 @@ const (
 
 type LambdaExecutor struct {
 	apiContainerSocket        string
-	logLevelStr               string
 	serializedCustomParamsStr string
 	configurator              LambdaConfigurator
 }
 
-func NewLambdaExecutor(apiContainerSocket string, logLevelStr string, serializedCustomParamsStr string, configurator LambdaConfigurator) *LambdaExecutor {
-	return &LambdaExecutor{apiContainerSocket: apiContainerSocket, logLevelStr: logLevelStr, serializedCustomParamsStr: serializedCustomParamsStr, configurator: configurator}
+func NewLambdaExecutor(apiContainerSocket string, serializedCustomParamsStr string, configurator LambdaConfigurator) *LambdaExecutor {
+	return &LambdaExecutor{apiContainerSocket: apiContainerSocket, serializedCustomParamsStr: serializedCustomParamsStr, configurator: configurator}
 }
 
 func (executor LambdaExecutor) Run() error {
-	if err := executor.configurator.SetLogLevel(executor.logLevelStr); err != nil {
-		return stacktrace.Propagate(err, "An error occurred setting the loglevel before running the lambda")
-	}
 
 	lambda, err := executor.configurator.ParseParamsAndCreateLambda(executor.serializedCustomParamsStr)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred parsing the serialized custom params and creating the lambda")
 	}
 
-	var lambdaServiceClient kurtosis_lambda_rpc_api_bindings.LambdaServiceClient = nil
+	var networkCtx *networks.NetworkContext
 	if executor.apiContainerSocket != "" {
-		// TODO SECURITY: Use HTTPS to ensure we're connecting to the real Lamba API servers
-		conn, err := grpc.Dial(executor.apiContainerSocket, grpc.WithInsecure())
+		// TODO SECURITY: Use HTTPS to verify we're hitting the correct API container
+		conn, err := grpc.Dial(executor.apiContainerSocket , grpc.WithInsecure())
 		if err != nil {
-			return stacktrace.Propagate(
-				err,
-				"An error occurred creating a connection to the Kurtosis API server at '%v'",
-				executor.apiContainerSocket,
-			)
+			return stacktrace.Propagate(err, "An error occurred dialling the API container at '%v'", executor.apiContainerSocket)
 		}
-		defer conn.Close()
 
-		lambdaServiceClient = kurtosis_lambda_rpc_api_bindings.NewLambdaServiceClient(conn)
+		apiClient := kurtosis_core_rpc_api_bindings.NewApiContainerServiceClient(conn)
+		networkCtx = networks.NewNetworkContext(
+			apiClient,
+			map[services.FilesArtifactID]string{},
+			kurtosis_lambda_docker_api.ExecutionVolumeMountpoint,
+		)
 	}
 
-	lambdaServiceServer := NewLambdaServiceServer(lambda, lambdaServiceClient)
+	lambdaServiceServer := NewLambdaServiceServer(lambda, networkCtx)
 	lambdaServiceRegistrationFunc := func(grpcServer *grpc.Server) {
 		kurtosis_lambda_rpc_api_bindings.RegisterLambdaServiceServer(grpcServer, lambdaServiceServer)
 	}