+++ docs/changelog.md
@@ -1,4 +1,14 @@
 # TBD
+
+# 0.4.0
+### Removed
+* Removed the Lambda loglevel environment variable
+
+### Breaking Changes
+* Removed the Lambda loglevel environment variable as a firstclass concept, since that should be handled by the Lambda itself 
+    * Users depending on this variable should push loglevel-setting into the Lambda custom initialization params
+
+# 0.3.1
 ### Features
 * Add the set of Docker envvars as constants that can be used by both the API container (to send the constants) and Lambda (to receive the constants)
 * Added a Lambda custom params Docker environment variable, analogous to a constructor, so Lambdas can accept data upon creation that will modify their behaviour